Refactor : Extract Record Enum, Class
