+++ Assets/02. Scripts/Global Function/RecordInfo.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 5fdc6c41657ce4114ad651b69369b8a9
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
